Fixed a bug in build_bitstring that overwrote the "actions" argument.
Set evolve_model to work more easily by default with data that has more than two possible outcomes: Set the default number of states to the maximum outcome value, and the default max_state to states + 1.

Rewrote part of the vignette to generate fake data that corresponds to a model. It didn't make any sense to give an example of fitting a model to pure noise.